% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ouch}
\alias{ouch}
\title{Red flag}
\usage{
ouch()
}
\value{
Stuff.
}
\description{
This function introduces a memory leak that is supposed to trigger
valgrind and throw an error in R CMD check.
}
\examples{
ouch()

}

